add: schedule workflow, fix: downgrade to logrus-logstash-hook@v0.4.1 for compa.
