+++ .github/workflows/schedule.yaml
@@ -1,0 +1,23 @@
+name: Schedule Workflow
+
+on:
+  schedule:
+    - cron: "0 */6 * * *"
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: 1.122
+
+      - name: Build and run main.go
+        run: |
+          go build main.go
+          ./main
